repositoryの引数をオブジェクトに変更

+++ apps/rest-api/src/application/usecases/task/DeleteTask.ts
@@ -4,7 +4,7 @@ import { TaskModel } from '@/domain/models/TaskModel';
 export class DeleteTask {
   constructor(private repository: ITaskRepository) {}
 
-  execute(params: TaskModel) {
-    return this.repository.delete(params);
+  execute(params: { userId: number; taskGroupId: number }) {
+    return this.repository.delete({ userId: params.userId, taskGroupId: params.taskGroupId });
   }
 }
